refactor: remove duplication with formStatus
